notification gets own modal
